service-python/cachingData/requirements.txt
Flask==0.12.1
Flask-SQLAlchemy==2.2
Mako==1.0.6
packaging==16.8
gunicorn==19.9.0
psycopg2==2.7.7
apscheduler==3.7.0
numpy==1.16.6
pymongo==3.11.0
python-dotenv==0.10.3
gunicorn==19.9.0
boto3==1.17.18